pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codesight"
version = "0.3.1"
description = "Semantic security verification CLI for scanner alerts and AI-written code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "CodeSight Maintainers"},
]
keywords = [
    "security", "semantic-security", "sarif", "code-review", "ai", "llm",
    "static-analysis", "vulnerability-detection", "cwe", "owasp",
    "cli", "devtools", "devsecops",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "anthropic>=0.40",
    "google-auth>=2.29",
    "rich>=13.7",
    "questionary>=2.0",
    "keyring>=24",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "build",
    "twine",
]

[project.scripts]
codesight = "codesight.cli:main"

[project.urls]
Homepage = "https://codesight.is-a.dev"
Documentation = "https://codesight.is-a.dev/docs"
Repository = "https://github.com/AvixoSec/codesight"
Issues = "https://github.com/AvixoSec/codesight/issues"
Changelog = "https://github.com/AvixoSec/codesight/releases"

[tool.setuptools.packages.find]
include = ["codesight*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
